Fix for crash with the following steps: 1. Click "Remove all settings" from the video screensaver Screensaver settings window 2. Open the video screensaver Screensaver settings window again 3. Click "Remove all settings" again Added check if key exists before deleting it

+++ VideoScreensaver/PreferenceManager.cs
@@ -25,7 +25,12 @@ namespace VideoScreensaver {
         public static void RemoveRegistryKeys()
         {
             RegistryKey software = Registry.CurrentUser.CreateSubKey("Software");
-            software.DeleteSubKeyTree(BASE_KEY);
+            var key = software.OpenSubKey(BASE_KEY);
+            if (key != null)
+            {
+                key.Close();
+                software.DeleteSubKeyTree(BASE_KEY);
+            }
         }
 
         public static List<String> ReadVideoSettings() {
